# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: WaCalls API
  version: "1.0.0"
  description: |
    API HTTP do **WaCalls** — chamadas de voz do WhatsApp pelo navegador.

    ## Autenticação
    Não há login/senha. Todas as rotas aceitam o header **`X-Client-Id`** (qualquer
    identificador estável), usado para marcar o "dono" das chamadas e filtrar eventos.

    > ⚠️ A API inicia/controla as chamadas (sinalização), mas o **áudio** exige um cliente
    > WebRTC (navegador) com captura de microfone — não é possível só com chamadas HTTP.
servers:
  - url: /
    description: Mesma origem do painel
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key definida em WACALLS_API_KEY (vazia = sem autenticação)"
  parameters:
    ClientId:
      name: X-Client-Id
      in: header
      required: false
      schema: { type: string, example: meuapp }
      description: Identificador do cliente (dono da chamada / filtro de eventos).
    Sid:
      name: sid
      in: path
      required: true
      schema: { type: string }
      description: ID da sessão (conta WhatsApp).
    CallId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: ID da chamada.
  schemas:
    Session:
      type: object
      properties:
        id: { type: string, example: 6458cca60a17c0cefa71de9da9dc92a2 }
        name: { type: string, example: WhatsApp }
        jid: { type: string, example: "556192264108:13@s.whatsapp.net" }
        state: { type: string, enum: [connecting, qr, open, logged_out, passkey_required], example: open }
        paired: { type: boolean, example: true }
    StartCallRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: "5561999999999", description: "DDI+DDD+número" }
        duration_ms: { type: integer, example: 300000 }
        record: { type: boolean, example: false }
paths:
...
  /api/sessions:
    get:
      tags: [Sessões]
      summary: Lista as contas
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
    post:
      tags: [Sessões]
      summary: Cria uma conta
      parameters: [{ $ref: "#/components/parameters/ClientId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "Minha conta" }
      responses:
        "200":
          description: Criada
          content:
            application/json:
              schema:
                type: object
                properties: { id: { type: string } }
  /api/sessions/{sid}:
    delete:
      tags: [Sessões]
      summary: Remove a conta
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Removida }
  /api/sessions/{sid}/pair:
    post:
      tags: [Sessões]
      summary: Inicia o pareamento (QR sai via SSE)
      description: Dispara a geração do QR Code; acompanhe os eventos `session-qr` / `auth-state` em `/api/events`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Pareamento iniciado }
  /api/sessions/{sid}/logout:
    post:
      tags: [Sessões]
      summary: Desconecta a conta (mantém a sessão)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Desconectada }
  /api/sessions/{sid}/import:
    post:
      tags: [Sessões]
      summary: Importa uma sessão a partir de um arquivo (wa-web dump, Baileys creds.json ou .zip)
      description: |
        multipart/form-data com o campo `file`. Substitui o device da sessão pelo importado
        e inicia a conexão; acompanhe via `/api/sessions/{sid}/importStatus` ou o evento `auth-state`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: Importada; conexão iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  jid: { type: string }
                  message: { type: string }
        "400": { description: Arquivo/formato inválido ou credenciais incompletas }
        "409": { description: Já existe uma importação em andamento para esta sessão }
  /api/sessions/{sid}/importRaw:
    post:
      tags: [Sessões]
      summary: Importa uma sessão a partir de JSON bruto (wa-web dump ou `{ creds, keys }` no formato Baileys)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                creds: { type: object }
                keys: { type: object }
      responses:
        "200":
          description: Importada; conexão iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  jid: { type: string }
                  message: { type: string }
        "400": { description: JSON/formato inválido ou credenciais incompletas }
        "409": { description: Já existe uma importação em andamento para esta sessão }
  /api/sessions/{sid}/export:
    get:
      tags: [Sessões]
      summary: Exporta as credenciais da sessão pareada (formato Baileys creds.json)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
        - name: format
          in: query
          schema: { type: string, enum: [json, zip], default: json }
      responses:
        "200": { description: "Download: creds.json ou session-{sid}.zip" }
        "400": { description: Sessão sem JID (não pareada) }
        "404": { description: Device não encontrado no banco }
  /api/sessions/{sid}/importStatus:
    get:
      tags: [Sessões]
      summary: Status da última importação e da conexão da sessão
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  jid: { type: string }
                  connectionStatus: { type: string, enum: [disconnected, connecting, connected] }
                  importStatus: { type: string, enum: [importing, connecting, error] }
                  importMessage: { type: string }
  /api/sessions/{sid}/calls:
    get:
      tags: [Chamadas]
      summary: Quantas chamadas ativas a sessão tem (e o limite)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  active: { type: integer, example: 2 }
                  maxCallsPerSession: { type: integer, example: 8 }
    post:
      tags: [Chamadas]
      summary: Inicia uma chamada (faz o número de destino tocar)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartCallRequest" }
      responses:
        "200":
          description: Chamada iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  call:
                    type: object
                    properties: { callId: { type: string } }
        "503": { description: Conta não pareada }
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      tags: [Chamadas]
      summary: Atende uma chamada recebida
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200": { description: Atendida }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      tags: [Chamadas]
      summary: Rejeita uma chamada recebida
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200": { description: Rejeitada }
  /api/sessions/{sid}/calls/{id}:
    get:
      tags: [Chamadas]
      summary: Consulta o estado de uma chamada específica
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, example: "ABCDEF1234" }
                  sid: { type: string, example: "6458cca60a17c0cefa71de9da9dc92a2" }
                  peer: { type: string, example: "5561999999999@s.whatsapp.net" }
                  state:
                    type: string
                    enum: [active, held, transferring, ended, ringing, initiating, connecting]
                    example: active
                  direction:
                    type: string
                    enum: [incoming, outgoing]
                    example: outgoing
        "404": { description: Chamada não encontrada }
    delete:
      tags: [Chamadas]
      summary: Encerra a chamada
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "204": { description: Encerrada }
  /api/sessions/{sid}/calls/{id}/hold:
    put:
      tags: [Chamadas]
      summary: Coloca a chamada em espera (hold)
      description: |
        Pausa o áudio do browser → peer e começa a injetar um tom de espera (440 Hz)
        no stream SRTP para o peer. A chamada WhatsApp permanece ativa.
        Se `moh_url` não for informado, silêncio é enviado pelo keepalive interno.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                moh_url:
                  type: string
                  format: uri
                  description: URL do áudio de Music on Hold (opcional).
                  example: "https://example.com/hold.mp3"
      responses:
        "200":
          description: Chamada em hold
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  state: { type: string, example: held }
        "404": { description: Chamada não encontrada }
        "409": { description: Chamada não está em estado ativo }
  /api/sessions/{sid}/calls/{id}/unhold:
    put:
      tags: [Chamadas]
      summary: Remove o hold — retoma áudio normal
      description: |
        Para a injeção de Music on Hold e retoma o bridge WebRTC normal
        (microfone do browser → peer, peer → browser).
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: Chamada ativa novamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  state: { type: string, example: active }
        "404": { description: Chamada não encontrada }
        "409": { description: Chamada não está em estado held }
  /api/sessions/{sid}/calls/{id}/pickup:
    post:
      tags: [Chamadas]
      summary: Transfere chamada ativa entre agentes (multiatendimento)
      description: |
        Transferência de chamada entre agentes sem encerrar a chamada WhatsApp.

        O servidor resolve a sessão real da chamada via broker — o `{sid}` no path
        pode ser qualquer sessão válida (ex: a sessão do próprio agente B).
        A chamada **nunca muda de sessão**; apenas o bridge WebRTC é substituído.

        Fluxo:

        1. **Agente B** chama este endpoint: `POST /api/sessions/{qualquer-sid}/calls/{id}/pickup`
        2. O servidor localiza a sessão real da chamada e fecha o bridge WebRTC do Agente A.
        3. O evento SSE `call-picked-up` é emitido com o `sessionId` real da chamada.
        4. **Agente B** conecta o áudio usando o `sessionId` retornado:
           `POST /api/sessions/{sessionId}/calls/{id}/webrtc`

        Não é necessário colocar a chamada em hold antes. Funciona com chamadas ativas.

        Evento SSE emitido: `call-picked-up` (`id`, `sessionId`).
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: Bridge anterior fechado; agente pode conectar via /webrtc
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  state: { type: string, example: active }
                  sessionId: { type: string }
        "404": { description: Chamada não encontrada na sessão }
  /api/sessions/{sid}/calls/{id}/webrtc:
    post:
      tags: [Chamadas]
      summary: Conecta canal de áudio WebRTC do agente à chamada ativa
      description: |
        Abre (ou substitui) o bridge WebRTC entre o browser do agente e a chamada
        WhatsApp em curso. Se já existir um bridge ativo (outro agente conectado),
        ele é fechado e substituído pelo novo.

        Uso típico pós-pickup:

        1. Agente B chama `POST /api/sessions/{sid}/calls/{id}/pickup` → recebe `sessionId`.
        2. Agente B chama este endpoint com `sessionId` retornado no passo anterior.

        O servidor responde com o SDP answer; o browser completa a negociação ICE
        e o áudio começa a fluir.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sdp_offer]
              properties:
                sdp_offer:
                  type: string
                  description: SDP offer gerado pelo RTCPeerConnection do browser.
      responses:
        "200":
          description: Bridge WebRTC estabelecido
          content:
            application/json:
              schema:
                type: object
                properties:
                  sdp_answer:
                    type: string
                    description: SDP answer para completar a negociação no browser.
        "400": { description: sdp_offer ausente }
        "404": { description: Chamada não encontrada na sessão }
        "500": { description: Falha ao criar bridge WebRTC }
  /api/sessions/{sid}/calls/{id}/transfer:
    post:
      tags: [Chamadas]
      summary: Transferência cega com fallback automático
      description: |
        Sequência:
        1. Coloca `{id}` em hold automaticamente (injeta Music on Hold para o peer).
        2. Na sessão `target_sid`, inicia uma chamada sainte para o mesmo peer.
        3. Retorna imediatamente com o `transfer_call_id`.
        4. Monitora assincronamente:
           - **Peer atende** → encerra a chamada original; emite SSE `transfer-completed`.
           - **Peer rejeita / timeout (60 s)** → encerra a transferência, restaura unhold; emite SSE `transfer-failed`.
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/CallId" }
        - { $ref: "#/components/parameters/ClientId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_sid]
              properties:
                target_sid:
                  type: string
                  description: ID da sessão de destino (deve ser diferente de `{sid}` e estar pareada).
                  example: "abc123"
                moh_url:
                  type: string
                  format: uri
                  description: URL do áudio de Music on Hold enquanto aguarda o destino atender (opcional).
                  example: "https://example.com/hold.mp3"
      responses:
        "200":
          description: Transferência iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  original_call_id: { type: string }
                  transfer_call_id: { type: string }
                  target_sid: { type: string }
                  state: { type: string, example: transferring }
        "400": { description: Parâmetros inválidos ou target_sid igual ao sid de origem }
        "404": { description: Chamada ou sessão destino não encontrada }
        "409": { description: Chamada não está em estado ativo }
        "503": { description: Sessão destino não pareada }
  /api/sessions/{sid}/history:
    get:
      tags: [Eventos & Histórico]
      summary: Histórico de chamadas da conta (últimas 50)
      parameters:
        - { $ref: "#/components/parameters/Sid" }
        - { $ref: "#/components/parameters/ClientId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items: { type: object }
  /api/events:
    get:
      tags: [Eventos & Histórico]
      summary: Stream de eventos em tempo real (SSE)
      description: |
        Server-Sent Events. Tipos existentes: `session-list`, `session-qr`, `auth-state`,
        `call-list`, `call-status`, `call-ended`, `incoming`, `incoming-claimed`.

        `auth-state` pode trazer `state: "passkey_required"` quando a WhatsApp exige
        autenticação via passkey para a conta — pareamento por QR code não é suportado
        nesse caso.

        Tipos adicionados pelo sistema de hold/transferência:

        | Tipo | Campos | Descrição |
        |---|---|---|
        | `call-held` | `id`, `sessionId` | Chamada entrou em hold |
        | `call-unheld` | `id`, `sessionId` | Hold removido, áudio normal retomado |
        | `transfer-started` | `original_call_id`, `transfer_call_id`, `sessionId` | Transferência iniciada |
        | `transfer-completed` | `original_call_id`, `transfer_call_id`, `sessionId` | Peer atendeu; chamada original encerrada |
        | `transfer-failed` | `original_call_id`, `transfer_call_id`, `sessionId`, `reason` (`rejected`\|`timeout`) | Transferência falhou; unhold automático |
        | `call-picked-up` | `id`, `sessionId` | Bridge anterior fechado; novo agente pode conectar via /webrtc |
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
          description: Identificador do cliente.
      responses:
        "200":
          description: Stream text/event-stream
          content:
            text/event-stream:
              schema: { type: string }